Refactored to fix bug again
