+++ main.py
@@ -3,11 +3,10 @@
 import json
 from wpimath.geometry import Pose3d
 
-from apriltag import apriltag, apriltag_estimator
-from net_table import network_tables
-from camera import camera
-import constants
-
+from src.apriltag import apriltag, apriltag_estimator
+from src.net_table import network_tables
+from src.camera import camera
+from src import constants
 #  Import all JSON settings
 with open("../config/Constants.json", encoding="utf-8") as file:
     settings = json.load(file)
