+++ src/__init__.py
@@ -1,3 +1,1 @@
-""" All code for our stuff """
-import sys
-sys.path.append(".")
+""" Init file :( """
